+++ tests/test_purtest.R
@@ -1,7 +1,7 @@
-# Various run tests for purtest()
+# Various run tests for purtest() and hansi()
 
-# NB: p-values can differ slightly relative to .Rout.save file due to availabilty of package 'urca'
-#     for p-value approximation in individiual (A)DF-regressions.
+# NB: p-values can differ slightly relative to .Rout.save file due to availability of package 'urca'
+#     for p-value approximation in individual (A)DF-regressions.
 
 library(plm)
 data("Grunfeld", package = "plm")
@@ -134,3 +134,20 @@ summary(llc_int)
 
 llc_trend <- purtest(pG$value, test = "levinlin", exo = "trend", lags = 0, dfcor = FALSE)
 summary(llc_trend)
+
+
+## Simes Test for panels by Hanck
+hansi(llc)
+hansi(llc_int)
+hansi(llc_trend)
+hansi(purtest(pG$value, test = "Pm",        exo = "intercept", lags = 2, dfcor = TRUE))
+hansi(purtest(pG$value, test = "invnormal", exo = "intercept", lags = 2, dfcor = TRUE))
+hansi(purtest(pG$value, test = "logit",     exo = "intercept", lags = 2, dfcor = TRUE))
+
+hansi(purtest(inv ~ 1, data = Grunfeld, index = "firm", pmax = 4, test = "madwu"))
+
+hansi(b_unbal3)
+hansi(b_ztbar)
+hansi(b_ztbar_unbal)
+hansi(b_ztbar_unbal2)
+hansi(b_lag2_ztbar)